Work-around for a bug relating to 4GB process size limit on Windows.
SIGAR treats all performance counters as 32-bit integers, which breaks when
certain values grow too large.

This primarily affects our process memory usage reporting, but will likely
affect other metrics. A proper fix would require the modification of the
PERF_VAL macro to look at the counter size: if CounterSize == 4, the value
is a DWORD; if CounterSize == 8, the value is a ULONGLONG (64-bit).